Stub feature flag API endpoint in tests
