+++ lib/feature-flags.test.js
@@ -1,31 +1,9 @@
 "use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || function (mod) {
-    if (mod && mod.__esModule) return mod;
-    var result = {};
-    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
-    __setModuleDefault(result, mod);
-    return result;
-};
 var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-const github = __importStar(require("@actions/github"));
 const ava_1 = __importDefault(require("ava"));
-const sinon = __importStar(require("sinon"));
-const apiClient = __importStar(require("./api-client"));
 const feature_flags_1 = require("./feature-flags");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
@@ -40,24 +18,6 @@ const testApiDetails = {
     url: "https://github.com",
 };
 const testRepositoryNwo = (0, repository_1.parseRepositoryNwo)("github/example");
-function mockHttpRequests(responseStatusCode, flags) {
-    // Passing an auth token is required, so we just use a dummy value
-    const client = github.getOctokit("123");
-    const requestSpy = sinon.stub(client, "request");
-    const optInSpy = requestSpy.withArgs("GET /repos/:owner/:repo/code-scanning/codeql-action/features");
-    if (responseStatusCode < 300) {
-        optInSpy.resolves({
-            status: responseStatusCode,
-            data: flags,
-            headers: {},
-            url: "GET /repos/:owner/:repo/code-scanning/codeql-action/features",
-        });
-    }
-    else {
-        optInSpy.throws(new util_1.HTTPError("some error message", responseStatusCode));
-    }
-    sinon.stub(apiClient, "getApiClient").value(() => client);
-}
 const ALL_FEATURE_FLAGS_DISABLED_VARIANTS = [
     {
         description: "GHES",
@@ -85,7 +45,7 @@ for (const variant of ALL_FEATURE_FLAGS_DISABLED_VARIANTS) {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
         const loggedMessages = [];
         const featureFlags = new feature_flags_1.GitHubFeatureFlags({ type: util_1.GitHubVariant.DOTCOM }, testApiDetails, testRepositoryNwo, (0, testing_utils_1.getRecordingLogger)(loggedMessages));
-        mockHttpRequests(200, {});
+        (0, testing_utils_1.mockFeatureFlagApiEndpoint)(200, {});
         for (const flag of Object.values(feature_flags_1.FeatureFlag)) {
             t.assert((await featureFlags.getValue(flag)) === false);
         }
@@ -104,7 +64,7 @@ for (const variant of ALL_FEATURE_FLAGS_DISABLED_VARIANTS) {
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
         const featureFlags = new feature_flags_1.GitHubFeatureFlags({ type: util_1.GitHubVariant.DOTCOM }, testApiDetails, testRepositoryNwo, (0, logging_1.getRunnerLogger)(true));
-        mockHttpRequests(500, {});
+        (0, testing_utils_1.mockFeatureFlagApiEndpoint)(500, {});
         await t.throwsAsync(async () => featureFlags.preloadFeatureFlags(), {
             message: "Encountered an error while trying to load feature flags: Error: some error message",
         });
@@ -125,7 +85,7 @@ for (const featureFlag of FEATURE_FLAGS) {
                 expectedFeatureFlags[f] = false;
             }
             expectedFeatureFlags[featureFlag] = true;
-            mockHttpRequests(200, expectedFeatureFlags);
+            (0, testing_utils_1.mockFeatureFlagApiEndpoint)(200, expectedFeatureFlags);
             const actualFeatureFlags = {
                 database_uploads_enabled: await featureFlags.getValue(feature_flags_1.FeatureFlag.DatabaseUploadsEnabled),
                 ml_powered_queries_enabled: await featureFlags.getValue(feature_flags_1.FeatureFlag.MlPoweredQueriesEnabled),
