+++ lib/testing-utils.js
@@ -19,9 +19,12 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getRecordingLogger = exports.setupActionsVars = exports.setupTests = void 0;
+exports.mockFeatureFlagApiEndpoint = exports.getRecordingLogger = exports.setupActionsVars = exports.setupTests = void 0;
+const github = __importStar(require("@actions/github"));
 const sinon = __importStar(require("sinon"));
+const apiClient = __importStar(require("./api-client"));
 const CodeQL = __importStar(require("./codeql"));
+const util_1 = require("./util");
 function wrapOutput(context) {
     // Function signature taken from Socket.write.
     // Note there are two overloads:
@@ -113,4 +116,24 @@ function getRecordingLogger(messages) {
     };
 }
 exports.getRecordingLogger = getRecordingLogger;
+/** Mock the HTTP request to the feature flags enablement API endpoint. */
+function mockFeatureFlagApiEndpoint(responseStatusCode, response) {
+    // Passing an auth token is required, so we just use a dummy value
+    const client = github.getOctokit("123");
+    const requestSpy = sinon.stub(client, "request");
+    const optInSpy = requestSpy.withArgs("GET /repos/:owner/:repo/code-scanning/codeql-action/features");
+    if (responseStatusCode < 300) {
+        optInSpy.resolves({
+            status: responseStatusCode,
+            data: response,
+            headers: {},
+            url: "GET /repos/:owner/:repo/code-scanning/codeql-action/features",
+        });
+    }
+    else {
+        optInSpy.throws(new util_1.HTTPError("some error message", responseStatusCode));
+    }
+    sinon.stub(apiClient, "getApiClient").value(() => client);
+}
+exports.mockFeatureFlagApiEndpoint = mockFeatureFlagApiEndpoint;
 //# sourceMappingURL=testing-utils.js.map
